Create new traits WithEvents, WithPolicies, WithRoutes
