+++ src/Illuminate/Foundation/Support/Providers/AuthServiceProvider.php
@@ -2,37 +2,16 @@
 
 namespace Illuminate\Foundation\Support\Providers;
 
-use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
 
 class AuthServiceProvider extends ServiceProvider
 {
+    use WithPolicies;
+
     /**
      * The policy mappings for the application.
      *
      * @var array<class-string, class-string>
      */
     protected $policies = [];
-
-    /**
-     * Register the application's policies.
-     *
-     * @return void
-     */
-    public function registerPolicies()
-    {
-        foreach ($this->policies() as $model => $policy) {
-            Gate::policy($model, $policy);
-        }
-    }
-
-    /**
-     * Get the policies defined on the provider.
-     *
-     * @return array<class-string, class-string>
-     */
-    public function policies()
-    {
-        return $this->policies;
-    }
 }
