+++ src/Illuminate/Foundation/Support/Providers/WithPolicies.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Illuminate\Foundation\Support\Providers;
+
+use Illuminate\Support\Facades\Gate;
+
+trait WithPolicies
+{
+    /**
+     * Register the application's policies.
+     *
+     * @return void
+     */
+    public function registerWithPolicies()
+    {
+        foreach ($this->policies() as $key => $value) {
+            Gate::policy($key, $value);
+        }
+    }
+
+    /**
+     * Get the policies defined on the provider.
+     *
+     * @return array<class-string, class-string>
+     */
+    public function policies()
+    {
+        return $this->policies ?? [];
+    }
+}
